requirements.txt
python-dotenv==1.0.0
pillow==10.0.0
requests==2.31.0
firebase-admin>=6.5.0
litellm
duckduckgo-search
gspread
google-api-python-client
google-auth
pytz
